Modify format of provenance output
Previously the `domain` could have been used to delegate which domain
the fact should be written to or the domain the entity is associated with.
Since this was ambiguous, the consumer is now responsible to delegating
where the fact should be written. The `domain` and `name` now identify
the entity.

References to other entities has been broken out `attrs` as `refs` to make
it easier for consumers to parse. Further, it will remove ambiguity for
nested documents in `attrs`.

A `batch` attribute has been defined which is used for tagging all entities
within a batch or snapshot with the same identifier. For stream processing
applications, this provides additional context of which entity came from
where.

In addition to the above, several bugs were fixed related to bad references.

Signed-off-by: Byron Ruth <b@devel.io>