Implement parser for rules files
This removes the need to maintain a copy of the rules in code and
update it whether the canonical source is updated.

The rules are fetched from the host repository on GitHub, parsed,
and used to rank the results.

This includes a breaking change in order to fetch the rule files from
GitHub. The assign-rank-to-issues command now requires the --token option
be supplied for authenticating with GitHub.

Signed-off-by: Byron Ruth <b@devel.io>