Re-open non-persistent issues across data cycles
Issues that have not been solved and are not persistent (i.e. no known
solution) must be re-opened across data cycles. The hope being that a
solution may be provided in the subsequent data cycle.

The new logic adds a check to see if the issue has the current data cycle
label and if not, adds it and re-opens the issue if closed.

Close #82