Ensure all persistent issues are copied
Prior to this fix only single-field issues were being copied because
it relied on a lookup against the data model. However there are some
multi-field issues. Since empty results are no longer added to the
template, the solution was simply to copy the issues from the previous
files and updating the version information.

Signed-off-by: Byron Ruth <b@devel.io>